requirements.txt
altair==5.5.0
antiorm==1.2.1
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cmake==4.0.2
colorama==0.4.6
cryptography==44.0.3
db==0.1.1
db-sqlite3==0.0.1
dlib==19.24.2
face-recognition==1.3.0
face-recognition-models==0.3.0
faiss-cpu==1.11.0
gitdb==4.0.12
GitPython==3.1.44
idna==3.10
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
MarkupSafe==3.0.2
narwhals==1.38.2
numpy==1.26.4
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
pillow==11.2.1
protobuf==6.30.2
pyaadhaar==2.0.2
pyarrow==20.0.0
pycparser==2.22
pydeck==0.9.1
pylibjpeg==2.0.1
pylibjpeg-openjpeg==2.4.0
python-dateutil==2.9.0.post0
pytz==2025.2
pyzbar==0.1.9
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
six==1.17.0
smmap==5.0.2
streamlit==1.45.0
tenacity==9.1.2
toml==0.10.2
tornado==6.4.2
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
watchdog==6.0.0
